# central-server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NEOPRO Central Server API
  description: |
    API RESTful pour la gestion centralisée des écrans publicitaires NEOPRO.

    ## Authentification

    L'API utilise l'authentification JWT. Obtenez un token via `/api/auth/login` et incluez-le dans le header `Authorization: Bearer <token>`.

    Les cookies HttpOnly sécurisés sont également supportés pour les applications web.

    ## MFA (Multi-Factor Authentication)

    Les comptes administrateurs peuvent activer l'authentification à deux facteurs (TOTP).
    Si MFA est activé, le login retournera `requireMfa: true` et nécessitera un code TOTP.

    ## Rate Limiting

    - Auth endpoints: 10 req/15 min
    - API générale: 100 req/min
    - Endpoints sensibles: 30 req/min
    - Uploads vidéo: 10/heure

    ## Codes d'erreur

    - `400` - Requête invalide
    - `401` - Non authentifié
    - `403` - Accès refusé (permissions insuffisantes)
    - `404` - Ressource non trouvée
    - `429` - Trop de requêtes (rate limit)
    - `500` - Erreur serveur
  version: 1.0.0
  contact:
    name: Support NEOPRO
    email: support@neopro.fr
servers:
  - url: https://api.neopro.fr
    description: Production
  - url: http://localhost:3001
    description: Développement
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: neopro_token
  schemas:
...
    CanaryDeployment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deploymentType:
          type: string
          enum: [content, update]
        resourceId:
          type: string
          format: uuid
        targetType:
          type: string
          enum: [site, group]
        targetId:
          type: string
          format: uuid
        config:
          $ref: '#/components/schemas/CanaryConfig'
        currentPhase:
          type: string
          enum: [canary, gradual, full, completed, failed, rolled_back]
        currentStep:
          type: integer
        metrics:
          $ref: '#/components/schemas/CanaryMetrics'
    CanaryConfig:
      type: object
      properties:
        canaryPercentage:
          type: integer
          default: 10
          description: Pourcentage de sites pour le canary initial
        gradualSteps:
          type: array
          items:
            type: integer
          default: [25, 50, 75, 100]
          description: Étapes de progression (%)
        stabilityPeriodMs:
          type: integer
          default: 1800000
          description: Temps d'observation entre phases (ms)
        successThreshold:
          type: integer
          default: 95
          description: Seuil de succès requis (%)
        autoAdvance:
          type: boolean
          default: true
          description: Avancement automatique
    CanaryMetrics:
      type: object
      properties:
        totalSites:
          type: integer
        deployedSites:
          type: integer
        successfulSites:
          type: integer
        failedSites:
          type: integer
        pendingSites:
          type: integer
        successRate:
          type: number
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/mfa/enable:
    post:
      tags: [MFA]
      summary: Activer le MFA après vérification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: Code TOTP à 6 chiffres
      responses:
        '200':
          description: MFA activé
        '400':
          description: Code invalide
  /api/mfa/disable:
    post:
      tags: [MFA]
      summary: Désactiver le MFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        '200':
          description: MFA désactivé
...
  /api/deployments:
    get:
      tags: [Deployments]
      summary: Liste des déploiements
      responses:
        '200':
          description: Liste des déploiements
    post:
      tags: [Deployments]
      summary: Créer un déploiement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [video_id, target_type, target_ids]
              properties:
                video_id:
                  type: string
                  format: uuid
                target_type:
                  type: string
                  enum: [site, group]
                target_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '201':
          description: Déploiement créé
  /api/canary/deployments:
    get:
      tags: [Canary]
      summary: Liste des déploiements canary actifs
      responses:
        '200':
          description: Liste des déploiements canary
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployments:
                    type: array
                    items:
                      $ref: '#/components/schemas/CanaryDeployment'
    post:
      tags: [Canary]
      summary: Créer un déploiement canary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deploymentType, resourceId, targetType, targetId]
              properties:
                deploymentType:
                  type: string
                  enum: [content, update]
                resourceId:
                  type: string
                  format: uuid
                targetType:
                  type: string
                  enum: [site, group]
                targetId:
                  type: string
                  format: uuid
                config:
                  $ref: '#/components/schemas/CanaryConfig'
      responses:
        '201':
          description: Déploiement canary créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  canarySites:
                    type: array
                    items:
                      type: string
  /api/canary/deployments/{id}/start:
    post:
      tags: [Canary]
      summary: Démarrer la phase canary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Phase canary démarrée
  /api/canary/deployments/{id}/advance:
    post:
      tags: [Canary]
      summary: Forcer l'avancement vers la phase suivante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Avancé vers la phase suivante
        '400':
          description: Impossible d'avancer (seuil non atteint)
  /api/canary/deployments/{id}/rollback:
    post:
      tags: [Canary]
      summary: Rollback du déploiement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Rollback effectué
...